make Checker class
